+++ ykrt/src/lib.rs
@@ -35,7 +35,7 @@ pub fn print_jit_state(state: &str) {
 }
 
 #[cfg(tracer_swt)]
-pub fn trace_basicblock(function_index: u32, block_index: u32) {
+pub fn trace_basicblock(function_index: usize, block_index: usize) {
     if mt::is_tracing() {
         trace::trace_basicblock(function_index, block_index)
     }
